+++ scripts/template-dapps-integration.sh
@@ -34,6 +34,9 @@ git clone https://github.com/multiversx/mx-template-dapp.git
 echo "cd mx-template-dapp..."
 cd mx-template-dapp
 
+echo "git checkout development..."
+git checkout development
+
 echo "Installing dependencies mx-template-dapp..."
 yarn install