+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added skip tests if PR is in draft](https://github.com/multiversx/mx-sdk-dapp/pull/1669)
 - [Refactored ledger helpers and add tests](https://github.com/multiversx/mx-sdk-dapp/pull/1668)
 - [Refactored `getLedgerProvider` and add helpers tests](https://github.com/multiversx/mx-sdk-dapp/pull/1667)
 - [Refactored `initApp` to use helpers & added tests](https://github.com/multiversx/mx-sdk-dapp/pull/1666)