+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[5.3.2](https://github.com/multiversx/mx-sdk-dapp/pull/1671)] - 2025-11-12
+
 - [Removed axios instance unused methods and added tests](https://github.com/multiversx/mx-sdk-dapp/pull/1670)
 - [Refactored ledger helpers and add tests](https://github.com/multiversx/mx-sdk-dapp/pull/1668)
 - [Refactored `getLedgerProvider` and add helpers tests](https://github.com/multiversx/mx-sdk-dapp/pull/1667)